Database insertion and retrieval implemented.  Edited CMake files for SQLite

+++ include/meme.h
@@ -2,8 +2,8 @@
 #include <string>
 
 struct Meme {
-  uint8_t meme_id;
-  uint8_t template_id;
+  uint32_t meme_id;
+  uint32_t template_id;
   std::string top_text;
   std::string bottom_text;
 };
